# docs/API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: Correspondence Board Game API
  description: REST API for managing asynchronous correspondence board games
  version: 1.0.0
  contact:
    name: Board Game API
    url: https://github.com/yourusername/correspondence-board-game
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-production-domain.com/api
    description: Production server
paths:
...
  /games:
    get:
      summary: Get all games
      description: Retrieve a list of all games, optionally filtered by status or game type
      tags:
        - Games
      parameters:
        - name: status
          in: query
          description: Filter games by status
          required: false
          schema:
            type: string
            enum: [waiting, active, completed, cancelled]
        - name: gameType
          in: query
          description: Filter games by game type
          required: false
          schema:
            type: string
            enum: [chess, checkers, hearts, solitaire]
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of games per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        '200':
          description: Successfully retrieved games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new game
      description: Create a new game with specified settings
      tags:
        - Games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - gameType
                - creatorId
              properties:
                name:
                  type: string
                  description: Name of the game
                  example: "Chess Match #1"
                gameType:
                  type: string
                  enum: [chess, checkers, hearts, solitaire]
                  description: Type of game to create
                  example: "chess"
                creatorId:
                  type: string
                  description: ID of the user creating the game
                  example: "user-123"
                settings:
                  type: object
                  description: Additional game settings (game-specific)
                  example: {}
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users:
    get:
      summary: Get all users (Admin)
      description: Retrieve a list of all registered users (admin feature)
      tags:
        - Users
        - Admin
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWithStats'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /game-types:
    get:
      summary: Get all game types
      description: Retrieve all supported game types from the plugin system
      tags:
        - Game Types
      responses:
        '200':
          description: Successfully retrieved game types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameType'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/purge:
    post:
      summary: Purge database (Admin)
      description: Delete all games, moves, and non-admin users from the database
      tags:
        - Admin
      responses:
        '200':
          description: Database purged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Database purged successfully"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Game:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the game
          example: "game-123"
        name:
          type: string
          description: Name of the game
          example: "Chess Match #1"
        gameType:
          type: string
          enum: [chess, checkers, hearts, solitaire]
          description: Type of game
          example: "chess"
        status:
          type: string
          enum: [waiting, active, completed, cancelled]
          description: Current status of the game
          example: "active"
        createdAt:
          type: string
          format: date-time
          description: When the game was created
          example: "2023-12-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the game was last updated
          example: "2023-12-01T11:30:00Z"
        players:
          type: array
          items:
            $ref: '#/components/schemas/GamePlayer'
          description: Players in the game
        maxPlayers:
          type: integer
          description: Maximum number of players for this game type
          example: 2
        currentPlayerId:
          type: string
          description: ID of the player whose turn it is
          example: "user-123"
          nullable: true
        moveCount:
          type: integer
          description: Total number of moves made in the game
          example: 15
        boardState:
          type: object
          description: Current state of the game board
          example: {}
        moves:
          type: array
          items:
            $ref: '#/components/schemas/GameMove'
          description: Move history (optional, included in some responses)
    GamePlayer:
      type: object
      properties:
        userId:
          type: string
          description: Unique identifier of the user
          example: "user-123"
        username:
          type: string
          description: Username of the player
          example: "player1"
        color:
          type: string
          description: Player's color/side in the game
          example: "white"
          nullable: true
        joinedAt:
          type: string
          format: date-time
          description: When the player joined the game
          example: "2023-12-01T10:05:00Z"
    GameMove:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the move
          example: "move-123"
        gameId:
          type: string
          description: ID of the game this move belongs to
          example: "game-123"
        playerId:
          type: string
          description: ID of the player who made the move
          example: "user-123"
        player:
          $ref: '#/components/schemas/GamePlayer'
        move:
          type: object
          description: Game-specific move data
          example: {"from": "e2", "to": "e4"}
        moveNumber:
          type: integer
          description: Sequential number of the move in the game
          example: 1
        timestamp:
          type: string
          format: date-time
          description: When the move was made
          example: "2023-12-01T10:10:00Z"
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the user
          example: "user-123"
        username:
          type: string
          description: Username of the user
          example: "player1"
        email:
          type: string
          format: email
          description: Email address of the user
          example: "player1@example.com"
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: When the user account was created
          example: "2023-11-01T09:00:00Z"
    UserWithStats:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            lastActive:
              type: string
              format: date-time
              description: When the user was last active
              example: "2023-12-01T10:00:00Z"
              nullable: true
            gamesPlayed:
              type: integer
              description: Total number of games the user has played
              example: 5
...
    GameType:
      type: object
      properties:
        type:
          type: string
          description: Unique identifier for the game type
          example: "chess"
        name:
          type: string
          description: Display name of the game type
          example: "Chess"
        description:
          type: string
          description: Description of the game type
          example: "Classic chess game for two players"
        minPlayers:
          type: integer
          description: Minimum number of players required
          example: 2
        maxPlayers:
          type: integer
          description: Maximum number of players allowed
          example: 2
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Game not found"
        code:
          type: string
          description: Error code (optional)
          example: "GAME_NOT_FOUND"
        details:
          type: object
          description: Additional error details (optional)
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...